+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+## 0.1.0
+  - Data is now stored in file extended attributes
+  - Got rid of DB
+  - Config is in /etc/whatmade/config.cfg
+
+
 ## 0.0.1 (untested)
-
-Features:
-
   - initial version
   - daemonizing
   - autocleaning DB (already deleted files)